Move project components into project directory
